add declaration parsing
declarations of the form

`ident = term`

can now be defined which will be parsed into the symbol table. all
declarations must come at the top of the program, and all programs must
have a function for evaluation (declarations are optional)

the symbol table was updated to use variables directly, so that they
will be easier to change in the future. variables also now must
implement `Hash`, so care is needed in the future to ensure that the
Variable struct stays hashable.