add church numerals example
a number of problems were fixed with the evaluation: when an abstraction
is encountered, an attempt is made to simplify the body. when we
evaluate an application in which the left term is a variable that is not
in the symbol table, we now recurse into the right hand side of the
application instead of giving up.

the parser accepts more identifiers for variables: any string that does
not contain one of the following characters: `\t\n λ.=()`.

printing was fixed for applications with applications as the right term;
the right term must be wrapped in parentheses, as applications are
normally left associative.